+++ web/src/main.tsx
@@ -13,7 +13,6 @@ window.React = React;
 window.ReactDOM = ReactDOM;
 
 loadConfig().finally(async () => {
-  await import("./services/config/aws.config");
   const element = document.getElementById("root");
   if (!element) throw new Error("root element is not found");
