+++ web/src/services/config/passwordPolicy.ts
@@ -1,0 +1,26 @@
+export type PasswordPolicy = {
+  tooShort?: RegExp;
+  tooLong?: RegExp;
+  whitespace?: RegExp;
+  lowSecurity?: RegExp;
+  medSecurity?: RegExp;
+  highSecurity?: RegExp;
+};
+
+export function convertPasswordPolicy(passwordPolicy?: {
+  [key: string]: string;
+}): { [key: string]: RegExp | undefined } | undefined {
+  if (!passwordPolicy) return;
+  return Object.fromEntries(
+    Object.entries(passwordPolicy)
+      .map(([k, v]) => {
+        if (typeof v !== "string") return [k, undefined];
+        try {
+          return [k, new RegExp(v)];
+        } catch {
+          return [k, undefined];
+        }
+      })
+      .filter(i => !!i[1]),
+  );
+}
